+++ backend/src/users/users.service.ts
@@ -210,6 +210,7 @@ export class UsersService {
       },
     });
 
+    if (!pendingItem) return;
     this.pendingRepository.remove(pendingItem);
   }
